stark_pipeline/requirements.txt
requests
beautifulsoup4
pdfplumber
pillow
transformers
torch
doctr
